refactor(productSupplier): change dto files to productSupplier folder and create new utils
